refactor: extract the app-wide security setup into a security package #307
