+++ backend/src/main/java/com/occupi/security/KeycloakRealmRoleConverter.java
@@ -1,4 +1,4 @@
-package com.occupi.feature.authentication;
+package com.occupi.security;
 
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.core.GrantedAuthority;
