+++ backend/src/test/java/com/occupi/security/KeycloakRealmRoleConverterTest.java
@@ -1,4 +1,4 @@
-package com.occupi.feature.authentication;
+package com.occupi.security;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
